Adds thin deployer to this repo
It's a Spring Cloud feature, but Spring Cloud is the main consumer
of thin launcher right now, so it feels natural to give it a home
here.

Also added tests to assert that the jar files are actually thin
(<1MB) because they weren't in some cases, when the gradle
build was broken.