Fix mixed up file and http transports in dependency resolver
The file: transport wasn't working (because it was wired up
wrong in Guice), so local repository resolution was broken. There
should have been an error, but that's a separate problem. This
change just gets the Guice configuration right so that there
is actually a file transport available.

It's hard to write a unit test (hence the new tests are integration
tests against the samples).