Cope with multiple jar tasks in the Gradle plugin
Previously, if a project contained multiple jar tasks, the plugin
would attempt to create mutiple tasks named thinResolvePrepare and
thinResolve. This caused a failure as tasks within a project must
be uniquely named.

This commit updates the task creation logic to prefix the names of the
created tasks with the name of the jar task that triggered their
creation. For example, a task named "jar" triggers the creation of
the "jarThinResolvePrepare" and "jarThinResolve" tasks.