Make thin resolve task names unique
If there are multiple jar tasks then they each get their own
"thinResolve" tasks, so we need a naming convention to keep them
unique. This change keeps the old default name if the jas task is
"jar", and appends a suffix of the capitalized task name if not.

Fixes gh-29