Add support for maven.repo.local in the wrapper
The only thing the wrapper needs to know is where to find the
launcher jar and where to put it. There are 2 components to
the origin: the maven artifact coordinates, and the repo location.
If there is a local maven cache the user might have specified the
latter as maven.repo.local, and so it is courtesy to follow that
hint if available and streamline the downloading (in case the
launcher is already cached).

See gh-52