Add support for thin.repo in build plugins
Enables the thin launcher to be downloaded explicitly from a
different location. Note that as long as it is cached locally that
copy will be used first anyway. Clarified this in the README.

Also simplified the "simple" sample a bit so it doesn't use the
thin plugin (it might look like to need that to build an executable
jar, and you don't).

Closes gh-60