Add support for profiles in pom.xml
If user sets thin.profile and one of the Maven profiles matches the
name of one of the thin profiles, it would be kind of natural to
use it. That's what this change does.

Also ensures that Maven profiles in ~/.m2/settings.xml get applied
when reading the pom.xml (in addition to any thin.profile values).

Maven specific of course. Gradle users will have to generate a pom.xml
with the profiles to take advantage.

See #69.