Change (again) the implementation of dryrun
The problem with the old new strategy was that it didn't give you
a thin root that you could run offline - boms in particular were
missing from the thin root. It's better to let Maven copy everything
you need into the thin root, even if it might be a bit slower.

To get back the main feature of the old new strategy - preferring
local snapshots - we now copy local snapshots to the thin root
manually as long as thin.local.snapshots is not "false" (default
is "true"). This seems to work empirically, but it's a bit hard
to test. Still thinking about that.